requirements.txt
Flask==3.0.0
Flask-MySQLdb==2.0.0
Flask-Limiter==3.5.0
flask-cors==4.0.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
gunicorn
mysql-connector-python==8.0.33